Change CMake configuration options
For now, disable async logging until it is fixed (both the busywaiting
and the logging queue growing without bound.

Also automatically decide if we are fuzzish or not based on whether
this is a debug or release build. Same for sanitizers, except we also
disable sanitizers if rust is enabled.

Finally modify compiler flags slightly