Oops my bad, bring in support elsewhere in the emulator needed to handle UART interrupts
This was breaking Seb's logic since we were never calling his interrupt
pending function, and thus his update function that read characters
asynchronously was never being called either. This fixes that.

Also implements the Data Ready bit in the LSR in a similar way to the
one in the ISR with the pending bit. Fixes my uart test which assumed
both were implemented :)